requirements.txt
python==3.11.4
customtkinter==5.2.0
requests==2.31.0
Pillow==10.0.0